Add cello_reduce_type for double or quad reduce
The reduction type is "cello_reduce_type" is currently controlled by
in-code defines CELLO_REDUCE_TYPE_QUAD (default) or
CELLO_REDUCE_TYPE_DOUBLE in cello.hpp

The purpose of replacing the hard-coded long double reductions with
cello_reduce_type is to bypass a performance bug in the NVIDIA
compiler on Grace-Hopper which calls floating-point emulation routines
for 128-bit values. This can slow down simulations by a factor of
about four.

Not that MethodFluxCorrection still uses "long double" even if
cello_reduce_type is set for "double" to maintain accuracy of
conserved quantities.